pyproject.toml
[tool.poetry]
name = "campus-rater"
version = "0.1.0"
description = ""
authors = ["Vinícius Cristofari Barbosa <viniciusnypbarbosa@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.30"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
bcrypt = "^4.1.3"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
databases = "^0.9.0"
asyncio = "^3.4.3"
email-validator = "^2.1.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.2.2"
mypy = "^1.10.0"
isort = "^5.13.2"
black = "^24.4.2"
types-passlib = "^1.7.7.20240327"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
ignored-modules = ["alembic.context", "alembic.op", "sqlalchemy.func"]
ignore-paths= ["./app/db/migrations/versions/**"]
suggestion-mode='yes'